refactor: create useEvaluationPageLogic
